requirements.txt
boto3
openpyxl
pandas
redshift_connector
ipykernel
xlrd
pyarrow
fastparquet
python-dotenv
altair
streamlit